refactors file editing
